backend/requirements.txt
Flask
Flask-Cors
Flask-JWT-Extended
python-dotenv
pymongo
scikit-learn
tensorflow
Keras
google-generativeai
pandas
numpy
requests
bcrypt
pillow
shap
googletrans==4.0.0rc1